Server 2/requirements.txt
numpy
pandas
Flask
matplotlib
scikit-learn
torch
Pillow
torchvision
flask-cors
gunicorn
